Simplification dans SED. Correction pour cacher le nom du compte Correction import avec split

+++ picsou/picsou/src/main/java/org/designup/picsou/triggers/ImportedToNotImportedAccountTransactionTrigger.java
@@ -59,7 +59,8 @@ public class ImportedToNotImportedAccountTransactionTrigger implements ChangeSet
 
       public void visitDeletion(Key key, FieldValues previousValues) throws Exception {
         Integer savingsTransaction = previousValues.get(Transaction.NOT_IMPORTED_TRANSACTION);
-        if (savingsTransaction != null) {
+        // la transaction peut avoir ete detruite par si la seriesbudget a été aussi detruite cf trigger SeriesBudgetUpdateTransactionTrigger
+        if (savingsTransaction != null && repository.find(Key.create(Transaction.TYPE, savingsTransaction)) != null) {
           repository.delete(Key.create(Transaction.TYPE, savingsTransaction));
         }
       }
